refactor: InitProcess 단계별 컴포넌트화
